Remove legacy plugin support
The new plugin infrastructure was introduced in v1.0 with deprecated
support for legacy (i.e. v0) plugins.

This PR removes the legacy plugin support from the catalog in
preparation for v1.1.

To limit the scope of the PR and ease review burden, this PR does not
remove any actual v0 plugins or interfaces. That will be done in follow
up PRs.

Signed-off-by: Andrew Harding <aharding@vmware.com>