Remove deprecated agent key (#2493)
In 1.0, the agent transitioned to a multi-key manager. The disk plugin
changed the storage location for keys to support this. It maintained
forwards and backwards compatability with the deprecated key location
(and format).

For 1.1, the deprecation key no longer needs to be supported. To prevent
keeping a historic key on disk, this PR updates the disk plugin to
remove the deprecated key. We can remove this code in a future release.

Signed-off-by: Andrew Harding <aharding@vmware.com>