Upgrade grpc-go to latest
A unit test was depending on an exact string to be present in an error
returned by grpc-go. The text in that error condition changed with the
latest release. Change test to use grpc-go error status handlers and
check for the status code rather than the exact error text, which is not
guaranteed to be stable across releases.

Signed-off-by: Ryan Turner <turner@uber.com>