Fix ghostunnel-federation integration test (#2534)
There is a misalignment in the libs available in the socat image v.s.
the ghostunnel image (namely openssl). This is causing a breakage since
the mashup image (using the ghostunnel image as a base) does not have
these libs anymore. Since ghostunnel has no dependencies (outside of
musl) and socat has many, it makes more sense to use the socat image as
the base for the mashup image.

This change updates the Dockerfile to use the socat image as the base.

Signed-off-by: Andrew Harding <aharding@vmware.com>