Validate https_spiffe profile field trust domains (#2561)
Relationships using self-serving endpoints have a bundle specified to
bootstrap the relationship. In this kind of profile both the bundle and
endpoint SPIFFE ID must match the trust domain of relationship.

Relationships using non self-serving endpoints are assumed to receive
the bundle to validate the endpoint via seperate configuration. They
therefore do not specify the bundle. The endpoint SPIFFE ID must NOT
match the trust domain of the trust domain relationship, since it is
implied that the endpoint is being served over a different trust domain.

Relevant spec text here:
https://github.com/spiffe/spiffe/blob/main/standards/SPIFFE_Federation.md#5222-endpoint-parameters

I also fixed up the audit fields being produced for bundles which were
not namespaced and in the case of the trust domain ID, were overwriting
the field set for the trust domain ID of the relationship itself.

Signed-off-by: Andrew Harding <aharding@vmware.com>