Fix workload attestation for K8s 1.21+
Kubernetes 1.21 seems to use cgroups namespaces differently than 1.20.
This breaks assumptions in the K8s workload attestor about how the paths
of the cgroups are shaped which prevents the attestor from locating the
pod/container of the process under attestation.

This change relaxes the regular expression to not expect the cgroup
paths we use to identify the pod/container to be expressely under a
kubepods prefix.

Fixes: #2578

Signed-off-by: Andrew Harding <aharding@vmware.com>