Maybe fix segfault in tpmdevid tests
This change updates the go-tpm-tools dependency to 0.3.1, which includes
some fixes around lifetimes. It also fixes up the tests which were in
many instances closing the simulator down more than once, which
wasn't safe in the 0.2.1 version and is an error in the 0.3.1 version.

To accomplish this, the change adds an OpenTPM function to the simulator
which returns the required ReadWriter for the session but does a noop
for closing, instead of closing down the simulator when the session is
closed. The simulator lifetime is tied to the test instead of the
session.

This _might_ solve our random segmentation faults we've been getting on
the unit tests during CI/CD, but this issue is really hard to reproduce
locally so it is hard to tell definitively.

Signed-off-by: Andrew Harding <aharding@vmware.com>