Validate upstream signed CA chain
This change introduces a validation step in the server CA before
accepting an upstream signed intermediate which ensures that an SVID
signed with that chain is valid. This prevents two kinds of
misconfigurations: 1) the upstream authority did not return an
intermediate chain that chains back to the upstream authority provided
root, 2) the upstream authority produced a chain with invalid
constraints (e.g. inappropriate key usage for certificate signing,
missing CA bit, etc.)

Signed-off-by: Andrew Harding <aharding@vmware.com>