Update README badges
This PR adds the new SPIFFE badge indicating production readiness.

I also took the liberty of removing the coveralls badge, as the number
hasn't moved in years and I don't believe that the project has found
this metric to be very useful. Perhaps it can be made more useful, but
for now it seems like a good idea to just remove it. Happy to put it
        back if folks would prefer to keep it.

Signed-off-by: Evan Gilman <egilman@vmware.com>