Document profiling configurations for Server and Agent
Adds documentation for configuration parameters used to set up a pprof
endpoint that have been around for a long time.

The entire config tables got affected because the description for
`profiling_endpoint` exceeded the length of any other config
parameter description. The table is formatted in the Markdown in
a human-friendly manner where all the columns are aligned, which
resulted in some noise in the diff to keep that convention consistent.

Signed-off-by: Ryan Turner <turner@uber.com>