RFC: Delegated  Identity API implementation (#2481)
* Delegated Identity API

This commit introduces the API's logic and the related files:

1. package created on "pkg/agent/api/delegatedidentity/v1"
Implement new spire API that allows users to get SVID of other workloads
by passing the selectors of such processes. Two RPC were created
"SubscribeToX509SVIDs" and "SubscribeToX509Bundles".

The API is available on the admin socket (the same used for the
debug api). A new "authorized_delegates" field is added in the agent config
file to set the SPIFFE IDs of the callers that are able to use this API.
Admin and downstream SVIDs are not sent over this API.

2. Other changes:
- Creation of a single workload attestor to be used in both endpoints: standard
 and admin;
- Code to validate the SPIFFE ID list used in configuration file of the agent;
- Create an middleware log for admin endpoints to be used by
rpccontext.Logger();
- Validation on workload cache selectors before sending the response.

Co-authored-by: Mauricio Vásquez <mauricio@kinvolk.io>
Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>
Signed-off-by: Raphael Campos <raphael@accuknox.com>