Give stable ordering of federates-with field
The delegated identity API response populates a string slice from a set
of map keys. This leads to inconsistent ordering in the response. This
probably shouldn't matter to consumers, but it does impact the tests,
which are asserting the response contents. While sorting the fields in
the test could also work, it doesn't seem like a heavy lift to just sort
these fields when crafting the response. It may help minimize false
positives on downstream consumers who may be doing naive response
equivalence to detect if something has changed, for example.

Signed-off-by: Andrew Harding <aharding@vmware.com>