Introduce centralized CLI printer
This commit introduces a centralized CLI printer that supports printing
output in multiple formats. It (generally) follows ideas from @gabece
in #1354, but stops short of implementing it everywhere.

In the future, we can use the pretty print facilities afforded by the
printer exclusively. In the meantime however, where we need to keep
compatibility with existing output formatting, the CLI printer supports
setting a custom pretty printer which can then print the legacy
formatting.

Signed-off-by: Evan Gilman <egilman@vmware.com>