Configurable for agent svid ttl
SPIRE Server currently has the `default_svid_ttl` configurable, which
dictates the TTL that should be used when signing SVIDs where a TTL is
not otherwise specified.

One quirk of this configurable is that since the SVIDs that agents use
are not minted against a registration entry, the default SVID TTL is
used. Since an agent's SVID TTL also constrains how long it can be
offline or unavailable for, it is a fairly important number that most
users put thought into and directly configure.

Since it can be desirable to have agents that can survive long(er)
periods offline, while still having a low default TTL for workload
SVIDs, it makes sense to allow agent SVID TTLs to be separately
configurable.

This commit introduces an `agent_ttl_configurable` on SPIRE Server. If
configured, all agent SVIDs will be minted with this TTL. If not
configured, SPIRE will fall back to the default SVID TTL.

Signed-off-by: Evan Gilman <egilman@vmware.com>