Fix race in gcp nodeattestor tests
The "not configured" test is racy since the plugin checks for
configuration status before it does the first receive on the stream.
This means that when it returns an error the client may observe a
failure to send the initial payload instead of a failure on the first
receive.

To fix this, this change moves the loading of the configuration after
the initial payload is received. This makes the test deterministic and
should not impact production, since plugins are configured when
initialized and before first use.

Signed-off-by: Andrew Harding <aharding@vmware.com>