Stop deleting deprecated key file (#2671)
The disk KeyManager plugin was updated several releases ago to manage
multiple keys instead of a single key. The path on disk changed at this
time to improve on the name and to also support version rollback. New
keys were written to both files.

In the next release, the plugin changed to remove the old key file.

In this release, the plugin no longer needs to remove the old key file
which is assumed to have already been removed during the process of
upgrading SPIRE.

Signed-off-by: Andrew Harding <aharding@vmware.com>