Remove migration code handled by previous SPIRE servers
The supported SPIRE upgrade strategy is that the infrastructure is
upgraded to each minor version between the existing version and the
target. This means that migration code for schema versions handled by
the previous minor version can be opportunistically removed from the
code.

Since 1.1.X does not contain any migrations (in other words upgrading to
1.0.X would apply all migrations needed for 1.1.X), the code is left
with no migrations, which looks a little funny when you read it, but
the necessary scaffolding remains in place, with the same touch points
as before when adding a new migration.

For convenience, I also added a table showing the SPIRE and schema
changes over time.

Signed-off-by: Andrew Harding <aharding@vmware.com>