Fail agent path template execution if key missing (#2683)
The default behavior of the text/template library when a key is missing
is to render `<no value>`. This is problematic for agent path templates,
since the produced agent path would require percent encoding and would
therefore not be conformant with the SPIFFE ID specification.

Previously this default behavior has been allowed. However, as of SPIRE
1.0, these types of agent would have been rejected when attempting
to issue RPCs by the authorization layer, which rejects non-conformant
agent IDs (unless the safety valve was activated).

The safety valve is being removed to a subsequent PR. Prior to that,
this is a small change which:

- Unifies the path template parsing and execution
- Changes the missing key behavior to fail the execution
- Preserves "ensure leading slash" behavior on produced paths

Signed-off-by: Andrew Harding <aharding@vmware.com>