Preserve "ensure leading slash" behavior
This change restores "ensure leading slash" behavior to the API layer
when processing SPIFFEID messages. It further adds deprecation warnings
so that we can deprecate this unwanted behavior in SPIRE 1.3.

Unfortunately, threading a logger into the right spots caused some more
code churn that would otherwise be warranted for this kind of change.
This includes some shuffling of types to resolve import cycles.

Most of the grossness involved with this change will be removed for
SPIRE 1.3.

Signed-off-by: Andrew Harding <aharding@vmware.com>