Use official SVIDStore V1 interface (#2688)
We neglected to publish the SVIDStore V1 interface in the SDK when it
was introduced in version SPIRE 1.1.0. This means that for 1.1.x,
external plugins, if any, would have been written against the interface
checked into the SPIRE repository.

Unfortunately, the protobuf package name on the checked in version is
does not follow the SDK convention, so even if we fix and publish the
interface in the SDK, external plugins written against the "unofficial"
interface need to still work.

To accommodate this situation and not break existing plugins, this
change consumes the newly published interface from the SDK and renames
the version checked into SPIRE to v1unofficial. The unofficial version
is marked as deprecated in the catalog to guide consumers move to the
SDK version.

Since the protobuf package name remains the same (i.e.
spire.agent.svidstore), existing external plugins will continue to
function, though, if they update their SPIRE dependency, they will need
to fix up import paths to continue to build. Ideally they switch to the
official v1 version published in the SDK.

We can remove v1unofficial in the SPIRE 1.3 release.

I also fixed a small bug in the catalog code that prevented the
preferred version from being logged.

Signed-off-by: Andrew Harding <aharding@vmware.com>