Configure dependabot to automatically raise PRs for gomod dependency upgrades
Dependabot is a tool that detects when package dependencies in a project
are not referring to the latest versions and submits individual pull
requests in the project to upgrade each dependency to the latest
version. This should hopefully simplify the maintenance burden of
keeping dependencies in sync and prevent us from being subject to
vulnerabilities from old dependency versions.

The plugin is configured to run daily (Mon-Fri). I chose to have the
bot run at a time when all maintainers are generally expected to be
online and available to give all maintainers an opportunity to see
the PRs. For now, I have limited it to raising at most five open PRs
at any given time.

Signed-off-by: Ryan Turner <turner@uber.com>