Fix race condition in agent cache (#2699)
As reported by @loveyana in #2696, there is a race condition in the
agent cache when reading/writing the selector index. This is caused by
the getSelectorIndex function opportunistically populating the index
when queried for selectors. This function is however called by
goroutines under both read and write locks. If a reader calls
this function with a selector that does not exist in the index, then
the map is mutated, possibly while being concurrently accessed by other
readers, causing the race.

The fix is to have two functions for accessing the selector index with one
for reads and one for writes. The write function has the current
behavior and will mutate the map when needed to insert a selector into
the index. The read function will return a nil index for selectors that
don't exist into the index. This prevents the map from being mutated
during reads.

Signed-off-by: Andrew Harding <aharding@vmware.com>