Gracefully stop plugin servers
SPIRE currently calls `Stop` on the plugin servers. This immediately
closes the server and cancels outstanding RPCs. However, `Stop` does not
wait for RPCs to complete (i.e. respond to the cancelation).

This change updates SPIRE to instead call `GracefulStop`, which will
allow outstanding RPCs a chance to complete.

There is _some_ risk with this change, i.e. that a stalled plugin could
prevent the server from shutting down. However, this can (and should) be
mitigated by instituting timeouts on plugin requests, which we have
already added for all KeyManager requests. Other timeouts can be added
if necessary. In order to facilitate debugging this kind of scenario,
I've added logging when plugins are being unloaded.

There is an additional benefit to this change, which is that unit tests
can know that when a plugin is unloaded that there is no longer any
executing plugin code. Not having this feature makes it harder, in
certain circumstances, to construct stable (and data-race free) tests
(see #2711).

Signed-off-by: Andrew Harding <aharding@vmware.com>