Support paths in AWS instance profiles for AWS IID node attestation (#2825)
Referencing issue #2796, when an instance profile has a path in the
ARN (say
`arn:aws:iam::123412341234:instance-profile/some/path/profile-name`)
the AWS IID node attestor attempts to get the instance profile
information from AWS passing in both the path and the
name (`some/path/profile-name`). AWS, however, considers only the
name (`profile-name`) to be relevant and returns a ValidationError if
the path is included as the forward slashes in the path are considered
invalid. AWS IAM documentation indicates that profile names are simply
EC2 specific versions of role names which are guaranteed to be unique
regardless of path. The fix for the node attestor is to pull out only
the name of the instance profile from the string that was previously
being passed in.

Signed-off-by: Andrew Shugarts <andrew.shugarts@appian.com>