Add a basic feature flag system
This commit adds a basic system for feature flagging in SPIRE. As we
move towards a more continuous release process that allows us to safely
ship development features to users and contributors, feature flagging
will be a critical capability.

We have a long way to go - we still need to figure out how to deal with
feature flagging that affects our data models .. but for now, I think
this system should be good enough to get us most of the rest of the
cases.

Signed-off-by: Evan Gilman <egilman@vmware.com>