Update to go1.18.1
Also includes some changes to golangci-lint and some fixes for the new
linter version. The version bump here is required for go1.18.
Unfortunately not all of our currently linters support generics, so we
need to force the linter into go1.17 mode until they are addressed.

Also changes the way golangci-lint is installed to use the go toolchain
for the project to install it instead of prebuilt binaries that may be
built with an older version of Go (which can cause problems when certain
features are released, like generics).

Signed-off-by: Andrew Harding <aharding@vmware.com>