Update vault test fixtures
The vault test fixtures unfortunately have SHA1 hashes in the
certificates. This prevents us from upgrading to Go1.18, which disallows
SHA1 signatures by default.

This change regenerates the test fixture certificates. It also removes
the OpenSSL instructions for generation, which were brittle, in favor of
a go file that can be run to regenerate everything. It also changed the
directory layout to get rid of the structure used by the disk plugin,
where this was copied from.

Signed-off-by: Andrew Harding <aharding@vmware.com>