Fix racy dialing during endpoint tests
The endpoint tests used to set WithBlock when dialing the local address.
This was changed recently when named pipe support was added and the
utils.GRPCDialContext() helper was introduced (see #2976). However,
WithBlock() is required to prevent failures to issue the RPCs when the
server hasn't had time to listen yet.

This change changes GRPCDialContext to allow options to be passed. The
tests now pass WithBlock().

In addition, this change also removes a FailOnNonTempDialError() that
occasionally caused test failures for similar reasons. The test dialing
shouldn't fail under normal circumstances. If it does, test timeout on
the dial operation should handle that.

This change also renames a test file to be consistent with the other
file names.

Signed-off-by: Andrew Harding <aharding@vmware.com>