Add configurable for leader election resource lock type
The current hardcoded type, "configmaps", is deprecated in Kubernetes
v1.24+ and is disallowed from newer releases of client-go.

In order to enable safe migration, we need to give operators a choice of
the resource lock type.

This change introduces the configurable, as well as deprecating
"configmaps" with a warning.

In a future SPIRE release, we can remove support for "configmaps".

Signed-off-by: Andrew Harding <aharding@vmware.com>