Upgrade linter and turn on go1.18 linting
The bodyclose, structcheck, and unparam linters still lack support for
go1.18 and are disabled. However, those linters are more about
cleanliness than correctness.

nolintlint has been disabled to silence the nolint exemptions we have in
place for the above, disabled linters. We can re-enable it when the
linter gains full support for go1.18.

Signed-off-by: Andrew Harding <aharding@vmware.com>