Fix race in agent manager tests
The TestSVIDRotation test does the following:
1. Prepare a test SVID
2. Initialize and start the manager
3. Take a read lock on the rotation mutex
4. Advance time to signal the rotator that it should attempt
   rotation
5. Verify that the rotation does not happen while the lock is held.

Unfortunately, due to timing, it is possible for the newly started
manager to rotate the SVID before the read lock is held, which causes
failures during step 5.

This change fixes the test by initializing the manager, then holding the
lock before advancing time and running the manager.

There is a LOT of further cleanup that could take place in these tests,
like changing the way the assertions work with the rotation hook, which
is triggered when the rotator simply checks for rotation, not when the
rotation actually takes place. I didn't feel like shaking that up right
now so I went with the targeted fix.

Signed-off-by: Andrew Harding <aharding@vmware.com>