Issue #2700: Adds columns x509_ttl and jwt_ttl in registration_entries
This is step 1 of the plan to add support for distinct X509 and JWT TTLs on registration entries. This change adds two new columns to the database to support storing these values wth each registration entry.

Follow-up work in v1.5.0 will add support for these new values in the API and server config.

Signed-off-by: Dennis Gove <dgove1@bloomberg.net>