Log agent SPIFFE ID when TOFU check fails (#3211)
Attestors currently do not return to the agent or even log which agent
failed a TOFU check during attestation.

To prevent information leakage, we probably don't want to include the
agent ID in the error returned from the server. However, we should be at
least logging this information on the server to facilitate debugging.

Due to the current attestation flow, this is a little hard to log from
the core service handler. However, each nodeattestor that enforces TOFU
obtains that information from a common helper.

This change updates that helper to log a message that includes the
SPIFFE ID as well as a uniform PermissionDenied status to callers.

Fixes: #3205

Signed-off-by: Andrew Harding <aharding@vmware.com>