Replace go.uber.org/atomic with sync/atomic
In Go 1.19, the sync/atomic package gained new Int32, Int64, Uint32, and
Uint64 types. These provide very similar functionality to the types in
go.uber.org/atomic. We only use this package in one place for tests, so
remove the package dependency in favor of the standard library.

Replace usage of Uint32 type with Int32 since Uint32 in sync/atomic does
not provide a method to subtract values, and it is not possible to
declare a negative Uint32 in the call to Add(). This is a safer type
to use since subtracting from 0 with a Uint32 can cause overflow.

Signed-off-by: Ryan Turner <turner@uber.com>