Remove old server entrycache (#3350)
This old cache code was left around when the full in-memory cache was
added in case any issues arised from the in-memory cache. The in-memory
cache has been running stably since v0.12.0. Remove this old code that
is no longer referenced.

This also removes a dependency on github.com/hashicorp/golang-lru.

Signed-off-by: Ryan Turner <turner@uber.com>