Use test keys for keymanager and other tests
This change uses the available test keys instead of generating new keys
for the keymanager tests which speeds up the tests and avoids sporadic
timeout test failures we've observed in CI/CD, particularly when
generating 4096 bit RSA keys.

This change also updates the testkey package to no longer fail if a
test exceeds the number of pregenerated keys, instead opting to generate
the new key and save it to disk so it can be included in the PR
introducing the test change.

The agent fake keymanager was also updated to use test keys.

CI/CD has also been updated to fail the unit-test step if the git
repository is no longer clean, implying that there were possibly test
keys generated that were not part of the PR.

Additionally, to ensure that production key management flows are not
broken the real key generation routines are used when the NIGHTLY tests
are executing. To prevent spurious timeouts, the go test timeouts are
also removed.

Signed-off-by: Andrew Harding <aharding@vmware.com>