Remove X509-SVID-TTL field from datastore model (#3541)
In order to remove the x509_svid_ttl column in 1.6.0, and still support
downgrading, we need to remove the X509SvidTtl column from the GORM
model in 1.5.0. Otherwise entry creation/updates will fail after the
downgrade as the 1.5.x code will still try and set the column, which
won't exist anymore.

Removing the field from the model does mean that new 1.5.x deployments will
not have the x509_svid_ttl column. This shouldn't be problematic on
upgrade as long as the column removal migration we do in 1.6.0 is idempotent.

Signed-off-by: Andrew Harding <aharding@vmware.com>