Fix racy bundle client tests
This change fixes test failures in the bundle client package. The
failures were caused by non-goroutine safe manipulation of a map of
configurations used as a config source and also an errant assertion that
didn't account for production code behavior.

To fix the non-goroutine safe config source, a new type was introduced
that protected the underlying config map with a RW mutex.

The errant assertion assumed that only one bundle refresh would be
performed for a newly discovered trust domain. However, since the manual
refresh operation ends up kicking off a goroutine that will also
periodically refresh the bundle, under certain timing conditions, the
bundle is refreshed twice. The assertion was updated to ensure that the
bundle is updated at least once.

Fixes: #2840,#3401

Signed-off-by: Andrew Harding <aharding@vmware.com>