Add bundle_file_arn to plugin options for awssecret
This will allow us to specify addtional certificates that should
be added to the trust bundle when using non-self-signed certificates in
the cert_file_arn configuration field. Without this, it is only possible
to use the awssecret UpstreamAuthority with a self-signed cert.
Otherwise, you will end up with a partial chain when presented with the
trust bundle + client certificate chain.

This parameter is optional, and when omitted does not break existing
flows.

Signed-off-by: Jay Crumb <jay.crumb@unit21.ai>