Fix racy AttestAgent tests
The AttestAgent tests assert log contents at the end of the test. These
tests fail with race detection since it is very possible for the server
to still be emitting logs (e.g. audit logs) to the test log hook when
the test code does the assertion. This is due to the streaming nature of
the AttestAgent RPC.

The fix here is to move the code that asserts logs until after the test
server has been fully shut down and cleaned up. The server shutdown code
waits for outstanding RPCs to fully complete. This guarantees that the
server code is finished emitting logs to the log hook by the time the
contents are asserted.

Since this is the last test that used the temporary infrastructure to
skip certain known racy tests under CI/CD, that infrastructure is also
removed as part of this PR.

Fixes: #2841

Signed-off-by: Andrew Harding <aharding@vmware.com>