Document new bundle_file_path parameter
This new parameter is optional as long as you are using a self-signed CA
in cert_file_arn. Otherwise bundle_file_arn should include one or more
roots to be included in the trust bundle, one of which must be the end
of the chain for the certificate specified in cert_file_arn.

Signed-off-by: Jay Crumb <jay.crumb@unit21.ai>