Add support for specifying an intermediate CA in cert_file_arn
This is a pretty major refactor of the awssecret plugin which allows
specifying an intermediate CA in cert_file_arn and brings this plugin
much more inline with the disk upstreamauthority. In addition to
allowing intermediates, we are now much stricter about checking the
provided certificate, key, and bundle at plugin load time. On init, we
now check that either cert_file_arn is a single self-signed CA or that
it is not self signed and can be verified with the roots specified in
bundle_file_arn.

Signed-off-by: Jay Crumb <jay.crumb@unit21.ai>