Generate new certificates and keys on every test run
The static test keys and certificates that currently live in
testdata/keys/EC are brittle and won't pass additional validation added
to the plugin. The certificate is expired and if replaced we're just
kicking the can down the road.

This commit moves towards a setup similar to how the disk
upstreamauthority plugin works, with dynamically generated keys and
certificates that will never be out of date. Instead of static test
data, on each run we generate chains of keys and certificates with
generateTestData() which returns both a function that will wrap these
values in a fake secrets manager client and a struct containing the
keys/certs so that we can use the same values to assert against SPIRE's
returned X509CaChain and UpstreamX509Roots.

Signed-off-by: Jay Crumb <jay.crumb@unit21.ai>