Rename spire-oidc-provider scratch image to oidc-discovery-provider
In order to preserve the same naming as the alpine-based image published
to GCR, rename the spire-oidc-provider image to oidc-discovery-provider
to match the name referenced in documentation, examples, and the source
code.

The rename to spire-oidc-provider was previously done to convey that the
service is only usable with SPIRE and does not depend on SPIFFE APIs,
since the image is published under the `spiffe` image namespace.

Signed-off-by: Ryan Turner <turner@uber.com>