Update release documentation to reflect more currently followed practices (#3563)
* Update release documentation to reflect more currently followed practices

Much of the release documentation has fallen out of date. There are some
key differences to how we manage the project that are not accounted for
in the currently documented process:

- Correlation of PRs to releases is tracked with GitHub milestones
- Every release has its own dedicated release branch
- We are not triggering releases directly off of the HEAD of main branch
- We are post-1.0, where the digits of the version strings have
  different meanings than in pre-1.0 releases. In many cases we were
  using the term "major" where we should be using "minor", and "minor"
  when we should be using "patch".
- The maintainers follow an on-call style rotation, where duties are
  distributed across the maintainers depending on when they are on-call.

Out-of-scope in this PR is defining a process for major version
releases, e.g. 2.x, because we have not defined a conclusive process for
this yet post-1.0.

Signed-off-by: Ryan Turner <turner@uber.com>